+++ backend/src/main/java/com/ssafy/togeball/domain/chatroom/repository/CustomChatroomRepository.java
@@ -5,5 +5,5 @@ import com.ssafy.togeball.domain.chatroom.entity.RecruitChatroom;
 import java.util.List;
 
 public interface CustomChatroomRepository {
-    List<RecruitChatroom> findByTagIds(List<Long> tagIds);
+    List<RecruitChatroom> findByTagIds(List<Integer> tagIds);
 }
